create baby_name.py
This program helps the new couple to find the bay name. The number of characters in the baby name, name of the parents(father/mother), another name of the parent( mother/father)  should be explicitly given. Based on these arguments the program will randomly generate the 1000 names at a time and shows the most possible baby name based on the first letter of the parent.